Add automatic embed preview on HTML head
This is so when the link is shared on different platforms, it will have
a nice title/description/image. Right now this is just for the
individual posts, and automated to the max. List pages and other pages
don't have this embed.

The image chosen for the embed will be `embedimage` from the front
matter if available, otherwise skipped.