Implement img/figure CSS
A thing to notice is that if you use both the caption _and_ attribution
in the figure shortcode, it will by default show next to each other. I
could fix it at the template level, but for now I'm just using an extra
`<br>` in my attribute text in the Markdown.