Implement asides
I hate footnotes because half the time I forget what the footnote was
referencing, and the other half when you tap on a footnote, it has no
way to go back to the text it was referencing.

I'm not fully convinced on this inline style, on the margins would be
better, but I'd need a big restructure of the CSS/layout, in addition of
keeping my asides short, so for now inline it is.