Style pre/code scrollbar and fix right padding
Previously if the overflow was used, the padding right wouldn't be
respected, and so you'd scroll to the right, and it'd just end right at
the end of the text which is awkward. Fixed it with the `display:
inline-block` on the code element.

Also styled the scrollbar so that the background/foreground colours
match the code block colours, and made it so the radius also applies to
the scrollbar. This one I don't really like the look of but I can
tolerate it for now.