Match shadow and spacing figure CSS in code blocks
The spacing around the code blocks was quite different to the spacing
around the figure blocks. Likewise I also got the feedback to match the
shadow in code blocks, so right now the code blocks should look pretty
much the same as the figure blocks. Spacing will likely be a tiny bit
off in the bottom compared to figures, due to the figure spacing
depending on the figcaption block, but it's close enough that I'm okay
hardcoding it like this.