Automatically include width/height of figures
This makes it so the layout won't shift around while scrolling, while
still keeping the lazy-load behaviour so all the images aren't loaded
with the first request.

Sadly this is not an option for the `gifvideo` figure as Hugo simply
doesn't support getting the height and width of video files yet.