Implement fetchpriority in figure shortcode
This makes it so we can give hints to the browser to download certain
images immediately. For my articles that I'll have one front and center
main image, this is highly useful. As an example, by implementing this
in the `kindle-start.webp` image of the _Reverse engineering Bluetooth
on Amazon Kindle eReader_ article, the LCP time went down from 0.50s to
0.18s, with no other optimisations.